Update Scripts for 8A and 8B
The two scripts were missing prompts that dealt with running the tests
from a non-clean database.

Add multiwait to 8B in place of the 'are you adding' to compensate for the
division already being added.

Add a wait and write after the 'OK' section in 8A to account for a prompt
that was left out.  You should wait for 'PATIENT NOTIFICATION TITLE' and
answer with a return ('')